# frontend/packages/api-client/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomerRoleProperties": {
        "additionalProperties": false,
        "properties": {
          "paymentMethodCode": {
            "anyOf": [
              {
                "enum": [
                  "payment/paypal",
                  "payment/credit-card",
                  "payment/sepa-direct-debit",
                  "payment/bank-transfer",
                  "payment/invoice"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paymentmethodcode"
          },
          "roleStatusCode": {
            "default": "role/active",
            "enum": [
              "role/active",
              "role/inactive"
            ],
            "title": "Rolestatuscode",
            "type": "string"
          }
        },
        "title": "CustomerRoleProperties",
        "type": "object"
      },
      "CustomerRoleRequest": {
        "additionalProperties": false,
        "properties": {
          "properties": {
            "$ref": "#/components/schemas/CustomerRoleProperties"
          },
          "type": {
            "const": "person/customer",
            "default": "person/customer",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties"
        ],
        "title": "CustomerRoleRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Consistent error contract distinguishing validation/business/conflict/infrastructure outcomes.",
        "properties": {
          "detail": {
            "description": "Specific explanation for this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "title": {
            "description": "Short human-readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable machine-readable error category, e.g. 'not_found'.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Page_PersonResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[PersonResponse]",
        "type": "object"
      },
...
      "PersonCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "entityId": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "properties": {
            "$ref": "#/components/schemas/PersonProperties"
          }
        },
        "required": [
          "properties"
        ],
        "title": "PersonCreateRequest",
        "type": "object"
      },
      "PersonProperties": {
        "additionalProperties": false,
        "properties": {
          "addressLocalityName": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addresslocalityname"
          },
          "familyName": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Familyname",
            "type": "string"
          },
          "givenName": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Givenname",
            "type": "string"
          }
        },
        "required": [
          "givenName",
          "familyName"
        ],
        "title": "PersonProperties",
        "type": "object"
      },
      "PersonResponse": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "displayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Displaynamechain",
            "type": "array"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "Person",
            "default": "Person",
            "title": "Entitykind",
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/PersonProperties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          }
        },
        "required": [
          "entityId",
          "schemaVersion",
          "properties",
          "displayName",
          "displayNameChain"
        ],
        "title": "PersonResponse",
        "type": "object"
      },
      "PersonRoleCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "entityId": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "role": {
            "discriminator": {
              "mapping": {
                "person/customer": "#/components/schemas/CustomerRoleRequest",
                "person/traveller": "#/components/schemas/TravellerRoleRequest"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerRoleRequest"
              },
              {
                "$ref": "#/components/schemas/TravellerRoleRequest"
              }
            ],
            "title": "Role"
          }
        },
        "required": [
          "role"
        ],
        "title": "PersonRoleCreateRequest",
        "type": "object"
      },
      "PersonRoleResponse": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "displayNameChain": {
            "items": {
              "type": "string"
            },
            "title": "Displaynamechain",
            "type": "array"
          },
          "entityId": {
            "title": "Entityid",
            "type": "string"
          },
          "entityKind": {
            "const": "PersonRole",
            "default": "PersonRole",
            "title": "Entitykind",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerRoleProperties"
              },
              {
                "$ref": "#/components/schemas/TravellerRoleProperties"
              }
            ],
            "title": "Properties"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "type",
          "schemaVersion",
          "properties",
          "displayName",
          "displayNameChain"
        ],
        "title": "PersonRoleResponse",
        "type": "object"
      },
      "PersonRoleUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "role": {
            "discriminator": {
              "mapping": {
                "person/customer": "#/components/schemas/CustomerRoleRequest",
                "person/traveller": "#/components/schemas/TravellerRoleRequest"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerRoleRequest"
              },
              {
                "$ref": "#/components/schemas/TravellerRoleRequest"
              }
            ],
            "title": "Role"
          }
        },
        "required": [
          "role"
        ],
        "title": "PersonRoleUpdateRequest",
        "type": "object"
      },
      "PersonUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "properties": {
            "$ref": "#/components/schemas/PersonProperties"
          }
        },
        "required": [
          "properties"
        ],
        "title": "PersonUpdateRequest",
        "type": "object"
      },
...
      "TravellerRoleProperties": {
        "additionalProperties": false,
        "properties": {
          "roleStatusCode": {
            "default": "role/active",
            "enum": [
              "role/active",
              "role/inactive"
            ],
            "title": "Rolestatuscode",
            "type": "string"
          }
        },
        "title": "TravellerRoleProperties",
        "type": "object"
      },
      "TravellerRoleRequest": {
        "additionalProperties": false,
        "properties": {
          "properties": {
            "$ref": "#/components/schemas/TravellerRoleProperties"
          },
          "type": {
            "const": "person/traveller",
            "default": "person/traveller",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties"
        ],
        "title": "TravellerRoleRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Shared capability-oriented API reused by Customer and Staff Interaction.",
    "title": "Christopher Columbus Travel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/persons": {
      "get": {
        "operationId": "listPersons",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page's nextCursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's nextCursor.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PersonResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Persons",
        "tags": [
          "persons"
        ]
      },
      "post": {
        "operationId": "createPerson",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Person",
        "tags": [
          "persons"
        ]
      }
    },
    "/persons/{person_id}": {
      "delete": {
        "operationId": "deletePerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Delete Person",
        "tags": [
          "persons"
        ]
      },
      "get": {
        "operationId": "getPerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Person",
        "tags": [
          "persons"
        ]
      },
      "put": {
        "operationId": "updatePerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Person",
        "tags": [
          "persons"
        ]
      }
    },
    "/persons/{person_id}/roles": {
      "get": {
        "operationId": "listPersonRoles",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonRoleResponse"
                  },
                  "title": "Response Listpersonroles",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Person Roles",
        "tags": [
          "persons"
        ]
      },
      "post": {
        "operationId": "createPersonRole",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonRoleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Create Person Role",
        "tags": [
          "persons"
        ]
      }
    },
    "/persons/{person_id}/roles/{role_id}": {
      "delete": {
        "operationId": "deletePersonRole",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Delete Person Role",
        "tags": [
          "persons"
        ]
      },
      "get": {
        "operationId": "getPersonRole",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Person Role",
        "tags": [
          "persons"
        ]
      },
      "put": {
        "operationId": "updatePersonRole",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonRoleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Person Role",
        "tags": [
          "persons"
        ]
      }
    },
...
  }
}